removed quick-junit deps
