+++ Cargo.toml
@@ -178,7 +178,6 @@ phf = "0.11.2"
 pico-args = "0.5.0"
 prettyplease = "0.2.25"
 project-root = "0.2.2"
-quick-junit = "0.5.1"
 rayon = "1.10.0"
 regex = "1.11.1"
 ropey = "1.6.1"
